KEGG enrichment analysis

KEGG enrichment analysis

Usage

KEGG.enrichment(
  gene_list,
  Species = "Homo",
  genelist_idtype,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  ...
)

Arguments

gene_list

The list of genes, which is required to be ensembl_gene_id, external_gene_name, entrezgene_id one of the three gene types.

Species

Select the species, Homo or Mus.

genelist_idtype

There are ensembl_gene_id, external_gene_name, entrezgene_id types of genetic IDs to choose from.

pvalueCutoff

Cutoff value of pvalue.

qvalueCutoff

Cutoff value of qvalue.

minGSSize

Minimal size of genes annotated by Ontology term for testing.

maxGSSize

Maximal size of genes annotated for testing.

...

...

Value

A KEGG object.

Examples

extdatadir <- system.file("extdata", "Senescence-associated secretory phenotype.txt", package = "Linkage")
gene_list <- read.table(extdatadir)
KEGG <- KEGG.enrichment(gene_list = gene_list$V1, Species = "Homo", genelist_idtype = "external_gene_name")
#>   ensembl_gene_id external_gene_name entrezgene_id chromosome_name
#> 1 ENSG00000215601            TSPY24P          <NA>            chrY
#> 2 ENSG00000215603           ZNF92P1Y          <NA>            chrY
#> 3 ENSG00000231436            RBMY3AP          <NA>            chrY
#> 4 ENSG00000225878           SERBP1P2          <NA>            chrY
#> 5 ENSG00000236435            TSPY12P          <NA>            chrY
#> 6 ENSG00000278478                             <NA>            chrY
#>   start_position end_position
#> 1        8280198      8282209
#> 2        7853686      7854876
#> 3        9610571      9621276
#> 4        4801685      4802848
#> 5        7687739      7689548
#> 6       10070131     10070409
#>  [1] "650"   "22943" "654"   "1906"  "4318"  "4313"  "5069"  "4282"  "5228" 
#> [10] "1437"  "3489"  "3487"  "5054"  "3082"  "283"   "5328"  "11082" "1472" 
#> [19] "6405"  "3600"  "7424"  "3553"  "2246"  "3558"  "51384" "3490"  "3606" 
#> [28] "1435"  "5327"  "2252"  "51129" "2920"  "3627"  "6696"  "4319"  "4312" 
#> [37] "6364"  "3484"  "3486"  "2919"  "2921"  "3576"  "3596"  "2069"  "9235" 
#> [46] "10344" "6369"  "7425"  "3623"  "7124"  "284"   "3574"  "3084"  "6387" 
#> [55] "3569"  "4314"  "4321"  "80763" "1508"  "9518"  "4322"  "7422"  "718"  
#> [64] "2247"  "6354"  "6355"  "6357"  "6346"  "6347"  "6360"  "6348"  "6351" 
#> [73] "3479"  "58191" "7077"  "3485"  "3488"  "3552"  "6352"  "3586"  "5270" 
#> [82] "1604" 
#> Reading KEGG annotation online: "https://rest.kegg.jp/link/hsa/pathway"...
#> Reading KEGG annotation online: "https://rest.kegg.jp/list/pathway/hsa"...